+++ tests/dragon/test_request_dispatcher.py
@@ -26,6 +26,7 @@
 
 import gc
 import os
+import pathlib
 import subprocess as sp
 import time
 import typing as t
@@ -68,8 +69,12 @@ from smartsim._core.mli.infrastructure.storage.backbone_feature_store import (
 from smartsim._core.mli.infrastructure.storage.dragon_feature_store import (
     DragonFeatureStore,
 )
+from smartsim._core.mli.infrastructure.storage.feature_store import ModelKey, TensorKey
 from smartsim._core.mli.infrastructure.worker.torch_worker import TorchWorker
+from smartsim._core.mli.infrastructure.worker.worker import InferenceRequest
+from smartsim._core.mli.message_handler import MessageHandler
 from smartsim.log import get_logger
+from tests.dragon.utils.channel import FileSystemCommChannel
 
 logger = get_logger(__name__)
 
@@ -157,12 +162,13 @@ def test_request_dispatcher(
         processes.append(process)
         process.start()
         assert process.returncode is None, "The message pump failed to start"
-
+        time.sleep(5)
         # give dragon some time to populate the message queues
         for i in range(15):
             try:
                 request_dispatcher._on_iteration()
-                batch = request_dispatcher.task_queue.get(timeout=1.0)
+                batch = request_dispatcher.task_queue.get(timeout=10)
+                print(batch.__dict__)
                 break
             except Empty:
                 time.sleep(2)
@@ -173,7 +179,7 @@ def test_request_dispatcher(
                 raise exc
 
         assert batch is not None
-        assert batch.has_valid_requests
+        assert batch.has_callbacks
 
         model_key = batch.model_id.key
 
@@ -196,7 +202,7 @@ def test_request_dispatcher(
                     )
                 )
 
-            assert len(batch.requests) == 2
+            assert len(batch.callbacks) == 2
             assert batch.model_id.key == model_key
             assert model_key in request_dispatcher._queues
             assert model_key in request_dispatcher._active_queues
@@ -231,3 +237,78 @@ def test_request_dispatcher(
     # Try to remove the dispatcher and free the memory
     del request_dispatcher
     gc.collect()
+
+
+def test_request_batch(test_dir: str) -> None:
+    tensor_key = TensorKey(key="key", descriptor="desc1")
+    tensor_key2 = TensorKey(key="key2", descriptor="desc1")
+    output_key = TensorKey(key="key", descriptor="desc2")
+    output_key2 = TensorKey(key="key2", descriptor="desc2")
+    model_id1 = ModelKey(key="model key", descriptor="model desc")
+    model_id2 = ModelKey(key="model key2", descriptor="model desc")
+    tensor_desc = MessageHandler.build_tensor_descriptor("c", "float32", [1, 2])
+    req_batch_model_id = ModelKey(key="req key", descriptor="desc")
+
+    callback1 = FileSystemCommChannel(pathlib.Path(test_dir) / "callback1")
+    callback2 = FileSystemCommChannel(pathlib.Path(test_dir) / "callback2")
+    callback3 = FileSystemCommChannel(pathlib.Path(test_dir) / "callback3")
+
+    request1 = InferenceRequest(
+        model_key=model_id1,
+        callback=callback1,
+        raw_inputs=[b"input data"],
+        input_keys=[tensor_key],
+        input_meta=[tensor_desc],
+        output_keys=[output_key],
+        raw_model=b"model",
+        batch_size=0,
+    )
+
+    request2 = InferenceRequest(
+        model_key=model_id2,
+        callback=callback2,
+        raw_inputs=None,
+        input_keys=[tensor_key, tensor_key2],
+        input_meta=None,
+        output_keys=[output_key, output_key2],
+        raw_model=b"model",
+        batch_size=0,
+    )
+
+    request3 = InferenceRequest(
+        model_key=model_id2,
+        callback=callback3,
+        raw_inputs=None,
+        input_keys=[tensor_key, tensor_key2],
+        input_meta=[tensor_desc],
+        output_keys=None,
+        raw_model=b"model",
+        batch_size=0,
+    )
+
+    request_batch = RequestBatch.from_requests(
+        [request1, request2, request3], None, req_batch_model_id
+    )
+
+    print(request_batch.__dict__)
+    assert len(request_batch.callbacks) == 3
+    for callback in request_batch.callbacks:
+        assert isinstance(callback, FileSystemCommChannel)
+    assert len(request_batch.output_key_refs.keys()) == 2
+    assert request_batch.has_callbacks
+    assert request_batch.model_id == req_batch_model_id
+    assert request_batch.inputs == None
+    assert request_batch.raw_model == b"model"
+    assert request_batch.raw_inputs == [b"input data"]
+    assert request_batch.input_meta == [tensor_desc, tensor_desc]
+    assert request_batch.input_keys == [
+        tensor_key,
+        tensor_key,
+        tensor_key2,
+        tensor_key,
+        tensor_key2,
+    ]
+    assert request_batch.output_key_refs == {
+        callback1: [output_key],
+        callback2: [output_key, output_key2],
+    }